requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
plotly>=5.22.0
requests>=2.32.0
python-dotenv>=1.0.1
fredapi>=0.5.1
scikit-learn>=1.5.0
statsmodels>=0.14.2
pytest>=8.2.0
openpyxl>=3.1.4
fpdf2>=2.7.0
